Add datatype interface for attribute and parameter datatype queries
This collection of interfaces enables the user to query the datatype
of parameter and attributes. The interfaces include Basic, Simple,
Enumerated, Array, FixedRecord, VariantRecord and NA datatypes.
The main class that does the grunt work is the DatatypeRetrieval
class.

////////////////////////////////////////////////////////////////
Variation from the Java Interface
///////////////////////////////////////////////////////////////
The setter functionality for the Datatype interfaces.
To keep data security and meet polymorphism constraints for the
cpp compilation I have removed the setters from this data should
not be modified by the end -user.
Another change was the IEnumerator interface has been removed
The additional layer of abstraction was not required for the cpp
portion... I think.

////////////////////////////////////////////////////////////////
Variation from the Java Additions
///////////////////////////////////////////////////////////////
Beside additions to the Java Proxy class there was an additional
class added for encoding the FOM datatypes into an XML object.
Currently, this only exports the datatypes but should eventually
have the rest of the FOM added.