Add interface to enable attribute/parameter datatype queries
This collection of interfaces enables the user to query the datatype
of parameter and attributes. The interfaces include Basic, Simple,
Enumerated, Array, FixedRecord, VariantRecord and NA datatypes.
The main class that does the grunt work is the DatatypeRetrieval
class.

_______________________________________________________
Variation from the Java Interface

The setter functionality for the Datatype interfaces.
To keep data security and meet polymorphism constraints for the
cpp compilation I have removed the setters from this data should
not be modified by the end -user.

Another change was the IEnumerator interface has been removed
The additional layer of abstraction was not required for the cpp
portion... I think.

_______________________________________________________
Variation from the Java Additions

Beside additions to the Java Proxy class, there was an additional class added
for encoding the FOM datatypes into an XML object. Currently, this only
exports the datatypes but should eventually have the rest of the FOM added.

_______________________________________________________
Dependencies
For the FOM xml parsing the pugixml library has been used.
The .lib files and include have been added to the build process.

****
This will need to be checked to make sure it is correct. The paths
have been added to the visual studio projects that require them.
****